backend/Bgclub/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.44.0
django-cors-headers==3.5.0
django-htmlmin==0.11.0
django-rest-auth==0.9.5
django-rest-registration==0.5.6
djangorestframework==3.12.2
html5lib==1.1
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
typing==3.7.4.3
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
